+++ test/unit/gen_dnf/dnf_fetch_install_pkgs.py
@@ -20,7 +20,7 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import gen_class                    # pylint:disable=E0401,R0402,C0413
+import gen_dnf                      # pylint:disable=E0401,R0402,C0413
 import version                      # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
@@ -48,7 +48,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.dnf = gen_class.Dnf()
+        self.dnf = gen_dnf.Dnf()
 
     def test_fetch_install_pkgs(self):
 
@@ -60,7 +60,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        data = self.dnf.fetch_update_pkgs()
+        data = self.dnf.fetch_install_pkgs()
 
         self.assertTrue(
             isinstance(data, list) and