+++ test/unit/gen_dnf/dnf_fetch_repos.py
@@ -1,11 +1,11 @@
 # Classification (U)
 
-"""Program:  dnf_get_hostname.py
+"""Program:  dnf_fetch_repos.py
 
-    Description:  Unit testing of Dnf.get_hostname in gen_class.py.
+    Description:  Unit testing of Dnf.fetch_repos in gen_class.py.
 
     Usage:
-        test/unit/gen_class/dnf_get_hostname.py
+        test/unit/gen_class/dnf_fetch_repos.py
 
     Arguments:
 
@@ -16,12 +16,11 @@
 # Standard
 import sys
 import os
-import socket
 import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import gen_class                    # pylint:disable=E0401,R0402,C0413
+import gen_dnf                      # pylint:disable=E0401,R0402,C0413
 import version                      # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
@@ -35,7 +34,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_get_hostname
+        test_repos
 
     """
 
@@ -49,21 +48,19 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.host_name = socket.gethostname()
+        self.dnf = gen_dnf.Dnf()
 
-    def test_get_hostname(self):
+    def test_repos(self):
 
-        """Function:  test_get_hostname
+        """Function:  test_repos
 
-        Description:  Test get_hostname method.
+        Description:  Test get_all_repos method.
 
         Arguments:
 
         """
 
-        dnf = gen_class.Dnf()
-
-        self.assertEqual(dnf.get_hostname(), self.host_name)
+        self.assertIsInstance(self.dnf.fetch_repos(), list)
 
 
 if __name__ == "__main__":